+++ crates/rhdl-fpga/src/serial_bus/mod.rs
@@ -30,6 +30,7 @@ pub mod mipi_dbi_type_b;
 pub mod mipi_dbi_type_c;
 pub mod nand_flash_async;
 pub mod one_wire_master;
+pub mod parallel_port_centronics;
 pub mod ps2_keyboard;
 pub mod ps2_mouse;
 pub mod rs485_master;